convert youtube-player tests to ts
